requirements.txt
numpy
pandas
flask
psycopg2-binary